Update tests
Remove old tests, move new tests to main directory, use more consistent module
naming, remove old stuff, related to two inputs only
