+++ benchmark/lib/beamchmark/h264_pipeline.exs
@@ -43,7 +43,7 @@ defmodule H264PipelineBeamchmark do
 
   alias Membrane.RawVideo
   alias Membrane.VideoCompositor.Implementations
-  alias Membrane.VideoCompositor.Utility
+  alias Membrane.VideoCompositor.Testing.Utility
 
   @spec benchmark(Implementations.implementation_t()) :: :ok
   def benchmark(implementation) do
