+++ benchmark/lib/pipelines/pipeline_h264_video.ex
@@ -18,7 +18,7 @@ defmodule Membrane.VideoCompositor.Benchmark.Pipeline.H264 do
   """
   @impl true
   def handle_init(options) do
-    decoder = Membrane.VideoCompositor.Benchmark.Pipeline.H264.TwoInputsParser
+    decoder = Membrane.VideoCompositor.Benchmark.Pipeline.H264.ParserDecoder
     encoder = Membrane.H264.FFmpeg.Encoder
 
     options = Map.put(options, :decoder, decoder)
