--- pipeline/lib/compose_two_inputs.ex
@@ -1,88 +1,0 @@
-defmodule Membrane.VideoCompositor.Pipeline.ComposeTwoInputs do
-  @moduledoc """
-  Universal pipeline for testing simple composing of two videos, by placing one above the other.
-  It loads two videos from the `options.first_video_path` and `options.second_video_path` files/src elements,
-  parses them using `options.decoder`, feeds VideoCompositor, encodes result (or simply pass by as RawVideo
-  if no `options.encoder` is specified) and feed `options.output_path` file/sink element.
-  """
-
-  use Membrane.Pipeline
-  alias Membrane.VideoCompositor.Pipeline.Utility.NoOp
-
-  @doc """
-  handle_init(%{
-    paths: %{
-        first_video_path: String.t() | Membrane.Source,
-        second_video_path: String.t() | Membrane.Source,
-        output_path: Membrane.Sink.t() | Membrane.Sink
-      },
-      caps: RawVideo.t(),
-      implementation: :ffmpeg | :opengl | :nx,
-      compositor: Membrane.Filter.t()
-      decoder: Membrane.Filter.t() | nil,
-      encoder: Membrane.Filter.t() | nil,
-  })
-  """
-  @impl true
-  def handle_init(options) do
-    children = %{
-      src_1: get_src(options.paths.first_video_path),
-      src_2: get_src(options.paths.second_video_path),
-      decoder_1: get_decoder(options),
-      decoder_2: get_decoder(options),
-      compositor: options.compositor,
-      encoder: get_encoder(options),
-      sink: get_sink(options.paths.output_path)
-    }
-
-    links = [
-      link(:src_1)
-      |> to(:decoder_1)
-      |> via_in(:first_input)
-      |> to(:compositor),
-      link(:src_2)
-      |> to(:decoder_2)
-      |> via_in(:second_input)
-      |> to(:compositor),
-      link(:compositor)
-      |> to(:encoder)
-      |> to(:sink)
-    ]
-
-    {{:ok, [spec: %ParentSpec{children: children, links: links}, playback: :playing]}, %{}}
-  end
-
-  defp get_src(input_path) when is_binary(input_path) do
-    %Membrane.File.Source{location: input_path}
-  end
-
-  defp get_src(src) when not is_nil(src) do
-    src
-  end
-
-  defp get_sink(output_path) when is_binary(output_path) do
-    %Membrane.File.Sink{location: output_path}
-  end
-
-  defp get_sink(sink) when not is_nil(sink) do
-    sink
-  end
-
-  defp get_encoder(options) do
-    Map.get(options, :encoder) || NoOp
-  end
-
-  defp get_decoder(options) do
-    Map.get(options, :decoder) || NoOp
-  end
-
-  @impl true
-  def handle_element_end_of_stream({:sink, :input}, _context, state) do
-    {{:ok, [playback: :terminating]}, state}
-  end
-
-  @impl true
-  def handle_element_end_of_stream({_pad, _src}, _context, state) do
-    {:ok, state}
-  end
-end
