+++ test/composing_test.exs
@@ -1,4 +1,4 @@
-defmodule Membrane.VideoCompositor.ComposingTest do
+defmodule Membrane.VideoCompositor.Test.Composing do
   @moduledoc false
   use ExUnit.Case
 
@@ -7,10 +7,10 @@ defmodule Membrane.VideoCompositor.ComposingTest do
   alias Membrane.RawVideo
   alias Membrane.Testing.Pipeline, as: TestingPipeline
   alias Membrane.VideoCompositor.Implementations
-  alias Membrane.VideoCompositor.Test.Pipeline.Raw, as: PipelineRaw
-  alias Membrane.VideoCompositor.Utility, as: TestingUtility
+  alias Membrane.VideoCompositor.Testing.Pipeline.Raw, as: PipelineRaw
+  alias Membrane.VideoCompositor.Testing.Utility, as: TestingUtility
 
-  @filter_description "split[b], pad=iw:ih*2[src], [src][b]overlay=0:h"
+  @filter_description "split[b1], pad=iw:ih*2[a1], [a1][b1]overlay=0:h, split[b2], pad=iw*2:ih[a2], [a2][b2]overlay=w:0"
   @implementation Implementations.get_all_implementations()
 
   @hd_video %RawVideo{
@@ -39,26 +39,42 @@ defmodule Membrane.VideoCompositor.ComposingTest do
 
   @spec test_raw_composing(Membrane.RawVideo.t(), non_neg_integer(), atom, binary(), binary()) ::
           nil
-  defp test_raw_composing(caps, duration, implementation, tmp_dir, sub_dir_name) do
+  defp test_raw_composing(video_caps, duration, implementation, tmp_dir, sub_dir_name) do
     {input_path, output_path, reference_path} =
-      TestingUtility.prepare_testing_video(caps, duration, "raw", tmp_dir, sub_dir_name)
+      TestingUtility.prepare_testing_video(video_caps, duration, "raw", tmp_dir, sub_dir_name)
 
     :ok =
       TestingUtility.generate_raw_ffmpeg_reference(
         input_path,
-        caps,
+        video_caps,
         reference_path,
         @filter_description
       )
 
+    positions = [
+      {0, 0},
+      {video_caps.width, 0},
+      {0, video_caps.height},
+      {video_caps.width, video_caps.height}
+    ]
+
+    out_caps = %RawVideo{
+      width: video_caps.width * 2,
+      height: video_caps.height * 2,
+      framerate: video_caps.framerate,
+      pixel_format: video_caps.pixel_format,
+      aligned: video_caps.aligned
+    }
+
     options = %{
       paths: %{
-        first_video_path: input_path,
-        second_video_path: input_path,
+        input_paths: for(_i <- 1..4, do: input_path),
         output_path: output_path
       },
-      caps: caps,
-      implementation: implementation
+      in_caps: video_caps,
+      caps: out_caps,
+      implementation: implementation,
+      positions: positions
     }
 
     assert {:ok, pid} = TestingPipeline.start_link(module: PipelineRaw, custom_args: options)
