+++ test/support/mock/frame_composer.ex
@@ -1,4 +1,4 @@
-defmodule Membrane.VideoCompositor.Test.Mock.FrameComposer.MultipleInput do
+defmodule Membrane.VideoCompositor.Testing.Mock.FrameComposer do
   @moduledoc """
   Mock frame composer for multiple input videos. It supports string buffers and concatenates them as an output.
   """
