+++ test/support/pipelines/h264/parser_decoder.ex
@@ -1,4 +1,4 @@
-defmodule Membrane.VideoCompositor.Test.Pipeline.H264.TwoInputsParser do
+defmodule Membrane.VideoCompositor.Testing.Pipeline.H264.ParserDecoder do
   @moduledoc """
   Simple bin parsing and decoding H264 buffers into raw frames.
   """
